Fix < 256m region's map tile generation.

+++ OpenSim/Region/CoreModules/World/WorldMap/TexturedMapTileRenderer.cs
@@ -321,15 +321,15 @@ namespace OpenSim.Region.CoreModules.World.WorldMap
             float waterHeight = (float)settings.WaterHeight;
 
             ITerrainChannel heightmap = m_scene.RequestModuleInterface<ITerrainChannel>();
-            int sizeRatio = m_scene.RegionInfo.RegionSizeX / Constants.RegionSize;
-            for (int y = 0; y < m_scene.RegionInfo.RegionSizeY; y += sizeRatio)
+            float sizeRatio = (float)m_scene.RegionInfo.RegionSizeX / (float)Constants.RegionSize;
+            for (float y = 0; y < m_scene.RegionInfo.RegionSizeY; y += sizeRatio)
             {
                 float rowRatio = y / (m_scene.RegionInfo.RegionSizeY - 1); // 0 - 1, for interpolation
-                for (int x = 0; x < m_scene.RegionInfo.RegionSizeX; x += sizeRatio)
+                for (float x = 0; x < m_scene.RegionInfo.RegionSizeX; x += sizeRatio)
                 {
                     float columnRatio = x / (m_scene.RegionInfo.RegionSizeX - 1); // 0 - 1, for interpolation
 
-                    float heightvalue = getHeight (heightmap, x, y);
+                    float heightvalue = getHeight (heightmap, (int)x, (int)y);
 
                     if (heightvalue > waterHeight)
                     {
@@ -375,12 +375,12 @@ namespace OpenSim.Region.CoreModules.World.WorldMap
                         }
                         //get the data from the original image
                         Color hsvColor = hsv.toColor ();
-                        unsafeBMP.SetPixel (x / sizeRatio, ((m_scene.RegionInfo.RegionSizeY - 1) - y) / sizeRatio, hsvColor);
+                        unsafeBMP.SetPixel ((int)(x / sizeRatio),(int)(((m_scene.RegionInfo.RegionSizeY - 1) - y) / sizeRatio), hsvColor);
                     }
                     else
                     {
                         // We're under the water level with the terrain, so paint water instead of land
-                        unsafeBMP.SetPixel (x / sizeRatio, ((m_scene.RegionInfo.RegionSizeY - 1) - y) / sizeRatio, WATER_COLOR);
+                        unsafeBMP.SetPixel ((int)(x / sizeRatio), (int)(((m_scene.RegionInfo.RegionSizeY - 1) - y) / sizeRatio), WATER_COLOR);
                     }
                 }
             }
